# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Global Tongue API Documentation",
        "description": "Heirarchy of API | Courses -> Levels -> Units -> Lessons -> Exercises",
        "contact": {
            "email": "gemeelijah@gmail.com"
        },
        "license": {
            "name": "www.globaltongueedu.com",
            "url": "https://globaltongueedu.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/public/api/courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get Course List",
                "description": "Get Course List as Array",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Get Course List as Array"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Create New Course",
                "description": "Create New Course",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "English for Beginner"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Here are the biggest enterprise technology acquisitions of 2021 so far, in reverse chronological order."
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Beginner"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "free"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create New Course"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/v1/public/api/courses/view/search": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "All Courses - Publicly Accessible",
                "description": "Search All Courses - Publicly Accessible",
                "operationId": "search",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "perPage, eg; 20",
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search, eg; Test",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Test"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All Courses - Publicly Accessible"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/v1/public/api/courses/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Show Course Details",
                "description": "Show Course Details",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Course Details"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Update Course",
                "description": "Update Course",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "English for Beginner"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Here are the biggest enterprise technology acquisitions of 2021 so far, in reverse chronological order."
                                    },
                                    "level": {
                                        "type": "string",
                                        "example": "Beginner"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "free"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Course"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Delete Course",
                "description": "Delete Course",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Course"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "apiKey",
                "description": "Bearer token, Ex: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}